# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Community mapping API for trees, mangroves, and trash sites
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EntryCategory = typeof EntryCategory[keyof typeof EntryCategory];


export const EntryCategory = {
  tree_planted: 'tree_planted',
  mangrove_dying: 'mangrove_dying',
  trash_pile: 'trash_pile',
} as const;

export interface Entry {
  id: number;
  lat: number;
  lng: number;
  category: EntryCategory;
  /** @nullable */
  notes?: string | null;
  reporter: string;
  date: string;
}

export type EntryInputCategory = typeof EntryInputCategory[keyof typeof EntryInputCategory];


export const EntryInputCategory = {
  tree_planted: 'tree_planted',
  mangrove_dying: 'mangrove_dying',
  trash_pile: 'trash_pile',
} as const;

export interface EntryInput {
  /**
     * @minimum -90
     * @maximum 90
     */
  lat: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  lng: number;
  category: EntryInputCategory;
  notes?: string;
  /** @minLength 1 */
  reporter: string;
}

export interface EntryStats {
  tree_planted: number;
  mangrove_dying: number;
  trash_pile: number;
  total: number;
}

export interface ErrorResponse {
  error: string;
}