# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Community mapping API for trees, mangroves, and trash sites
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: entries
    description: Map entry operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /entries:
    get:
      operationId: listEntries
      tags: [entries]
      summary: List all map entries
      description: Returns all saved map markers
      responses:
        "200":
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entry"
    post:
      operationId: createEntry
      tags: [entries]
      summary: Create a new map entry
      description: Save a new marker to the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntryInput"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /entries/stats:
    get:
      operationId: getEntryStats
      tags: [entries]
      summary: Get entry statistics
      description: Returns counts of entries grouped by category
      responses:
        "200":
          description: Entry statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryStats"

  /entries/{id}:
    delete:
      operationId: deleteEntry
      tags: [entries]
      summary: Delete a map entry
      description: Remove a specific marker by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Entry:
      type: object
      required: [id, lat, lng, category, reporter, date]
      properties:
        id:
          type: integer
        lat:
          type: number
        lng:
          type: number
        category:
          type: string
          enum: [tree_planted, mangrove_dying, trash_pile]
        notes:
          type: ["string", "null"]
        reporter:
          type: string
        date:
          type: string

    EntryInput:
      type: object
      required: [lat, lng, category, reporter]
      properties:
        lat:
          type: number
          minimum: -90
          maximum: 90
        lng:
          type: number
          minimum: -180
          maximum: 180
        category:
          type: string
          enum: [tree_planted, mangrove_dying, trash_pile]
        notes:
          type: string
        reporter:
          type: string
          minLength: 1

    EntryStats:
      type: object
      required: [tree_planted, mangrove_dying, trash_pile, total]
      properties:
        tree_planted:
          type: integer
        mangrove_dying:
          type: integer
        trash_pile:
          type: integer
        total:
          type: integer

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string