# lib/api-zod/src/generated/types/entry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Community mapping API for trees, mangroves, and trash sites
 * OpenAPI spec version: 0.1.0
 */
import type { EntryCategory } from './entryCategory';

export interface Entry {
  id: number;
  lat: number;
  lng: number;
  category: EntryCategory;
  /** @nullable */
  notes?: string | null;
  reporter: string;
  date: string;
}