# lib/api-zod/src/generated/types/entryInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Community mapping API for trees, mangroves, and trash sites
 * OpenAPI spec version: 0.1.0
 */
import type { EntryInputCategory } from './entryInputCategory';

export interface EntryInput {
  /**
     * @minimum -90
     * @maximum 90
     */
  lat: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  lng: number;
  category: EntryInputCategory;
  notes?: string;
  /** @minLength 1 */
  reporter: string;
}