# lib/api-zod/src/generated/types/entryStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Community mapping API for trees, mangroves, and trash sites
 * OpenAPI spec version: 0.1.0
 */

export interface EntryStats {
  tree_planted: number;
  mangrove_dying: number;
  trash_pile: number;
  total: number;
}